+++ include/gf/ResourceManager.h
@@ -22,15 +22,18 @@
 #define GF_RESOURCE_MANAGER_H
 
 #include <functional>
+#include <future>
 #include <initializer_list>
 #include <map>
 #include <memory>
 #include <stdexcept>
 
 #include "AssetManager.h"
+#include "Image.h"
 #include "Font.h"
 #include "Portability.h"
 #include "Texture.h"
+#include "Window.h"
 
 namespace gf {
 #ifndef DOXYGEN_SHOULD_SKIP_THIS
@@ -87,7 +90,16 @@ inline namespace v1 {
      * @return A reference to the resource
      * @throw std::runtime_error If the resource is not found
      */
-    T& getResource(AssetManager& assetManager, const Path& filename) {
+    T& getResource(AssetManager& assets, Path filename) {
+      // try to find a known prefix
+      if (filename.is_absolute()) {
+        Path relativePath = assets.getRelativePath(filename);
+
+        if (!relativePath.empty()) {
+          filename = relativePath;
+        }
+      }
+
       std::size_t h = boost::filesystem::hash_value(filename);
 
       auto it = m_cache.find(h);
@@ -96,7 +108,7 @@ inline namespace v1 {
         return *it->second;
       }
 
-      Path absolutePath = assetManager.getAbsolutePath(filename);
+      Path absolutePath = assets.getAbsolutePath(filename);
 
       if (absolutePath.empty()) {
         throw std::runtime_error("Path not found");
@@ -141,6 +153,19 @@ inline namespace v1 {
      */
     ResourceManager(std::initializer_list<Path> paths);
 
+
+    /**
+     * @brief Get an image
+     *
+     * @param path A path to the image
+     * @return A reference to the image
+     * @throw std::runtime_error If the font is not found
+     */
+    Image& getImage(const Path& path) {
+      std::lock_guard<std::mutex> lock(m_mutex);
+      return m_images.getResource(*this, path);
+    }
+
     /**
      * @brief Get a texture
      *
@@ -149,6 +174,7 @@ inline namespace v1 {
      * @throw std::runtime_error If the texture is not found
      */
     Texture& getTexture(const Path& path) {
+      std::lock_guard<std::mutex> lock(m_mutex);
       return m_textures.getResource(*this, path);
     }
 
@@ -160,12 +186,15 @@ inline namespace v1 {
      * @throw std::runtime_error If the font is not found
      */
     Font& getFont(const Path& path) {
+      std::lock_guard<std::mutex> lock(m_mutex);
       return m_fonts.getResource(*this, path);
     }
 
   private:
+    ResourceCache<Image> m_images;
     ResourceCache<Texture> m_textures;
     ResourceCache<Font> m_fonts;
+    std::mutex m_mutex;
   };
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS