+++ library/graphics/Window.cc
@@ -104,16 +104,25 @@ inline namespace v1 {
 
   Window::Window(StringRef title, Vector2i size, Flags<WindowHints> hints)
   : m_window(nullptr)
-  , m_context(nullptr)
+  , m_mainContext(nullptr)
+  , m_sharedContext(nullptr)
   , m_shouldClose(false)
   , m_isFullscreen(false)
   , m_vao(0)
   {
     auto flags = getFlagsFromHints(hints);
     m_window = SDL_CreateWindow(title.getData(), SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, size.width, size.height, flags);
-    m_context = createContextFromWindow(m_window);
 
-    if (m_context != nullptr) {
+    SDL_GL_SetAttribute(SDL_GL_SHARE_WITH_CURRENT_CONTEXT, 1);
+    m_sharedContext = SDL_GL_CreateContext(m_window);
+    m_mainContext = createContextFromWindow(m_window);
+
+    if (m_sharedContext == nullptr) {
+      Log::error("Failed to create a shared context: %s\n", SDL_GetError());
+      m_sharedContext = nullptr;
+    }
+
+    if (m_mainContext != nullptr) {
       glCheck(glEnable(GL_BLEND));
       glCheck(glEnable(GL_SCISSOR_TEST));
       glCheck(glClear(GL_COLOR_BUFFER_BIT));
@@ -128,12 +137,16 @@ inline namespace v1 {
   }
 
   Window::~Window() {
-    if (m_context != nullptr) {
+    if (m_sharedContext != nullptr) {
+      SDL_GL_DeleteContext(m_sharedContext);
+    }
+
+    if (m_mainContext != nullptr) {
 #ifdef GF_OPENGL3
       glCheck(glBindVertexArray(0));
       glCheck(glDeleteVertexArrays(1, &m_vao));
 #endif
-      SDL_GL_DeleteContext(m_context);
+      SDL_GL_DeleteContext(m_mainContext);
     }
 
     if (m_window != nullptr) {
@@ -391,7 +404,7 @@ inline namespace v1 {
       return modifiers;
     }
 
-    bool translateEvent(Uint32 windowId, const SDL_Event *in, Event& out) {
+    bool translateEvent(Uint32 windowId, Vector2i size, const SDL_Event *in, Event& out) {
       switch (in->type) {
         case SDL_WINDOWEVENT:
           if (windowId != in->window.windowID) {
@@ -540,6 +553,27 @@ inline namespace v1 {
           std::strncpy(out.text.rune.data, in->text.text, Rune::Size);
           break;
 
+        case SDL_FINGERDOWN:
+          out.type = EventType::TouchBegan;
+          out.touch.finger = in->tfinger.fingerId;
+          out.touch.coords.x = static_cast<int>(in->tfinger.x * size.width);
+          out.touch.coords.y = static_cast<int>(in->tfinger.y * size.height);
+          break;
+
+        case SDL_FINGERMOTION:
+          out.type = EventType::TouchMoved;
+          out.touch.finger = in->tfinger.fingerId;
+          out.touch.coords.x = static_cast<int>(in->tfinger.x * size.width);
+          out.touch.coords.y = static_cast<int>(in->tfinger.y * size.height);
+          break;
+
+        case SDL_FINGERUP:
+          out.type = EventType::TouchEnded;
+          out.touch.finger = in->tfinger.fingerId;
+          out.touch.coords.x = static_cast<int>(in->tfinger.x * size.width);
+          out.touch.coords.y = static_cast<int>(in->tfinger.y * size.height);
+          break;
+
         default:
           return false;
       }
@@ -561,7 +595,7 @@ inline namespace v1 {
       if (status == 0) {
         return false;
       }
-    } while (!translateEvent(windowId, &ev, event));
+    } while (!translateEvent(windowId, getSize(), &ev, event));
 
     return true;
   }
@@ -578,7 +612,7 @@ inline namespace v1 {
       if (status == 0) {
         return false;
       }
-    } while (!translateEvent(windowId, &ev, event));
+    } while (!translateEvent(windowId, getSize(), &ev, event));
 
     return true;
   }
@@ -627,6 +661,14 @@ inline namespace v1 {
     }
 
     SDL_SetCursor(cursor.m_cursor);
+  }
+
+  void Window::attachGLContext() {
+    SDL_GL_MakeCurrent(m_window, m_sharedContext);
+  }
+
+  void Window::detachGLContext() {
+    SDL_GL_MakeCurrent(m_window, nullptr);
   }
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS